Remove store_stocks endpoint and complete Loyverse integration
- Remove all references to unsupported store_stocks endpoint
- Add GetCategories method to connector.Client for consistency
- Update sync intervals in docker-compose.yml (Master Data: 12h, Real-time: 2-8h)
- Implement complete sync functionality for customer, inventory, receipt syncs
- Fix JSON unmarshaling issues (points_earned from int to float64)
- Add comprehensive endpoint testing with proper error handling
- Simplify main.go structure and remove unused admin middleware
- Update config to support extended sync intervals for all endpoints

+++ integrations/loyverse/internal/models/loyverse.go
@@ -165,9 +165,9 @@ type Receipt struct {
 	Source         string     `json:"source"`
 	TotalMoney     float64    `json:"total_money"`
 	TotalTax       float64    `json:"total_tax"`
-	PointsEarned   int        `json:"points_earned"`
-	PointsDeducted int        `json:"points_deducted"`
-	PointsBalance  int        `json:"points_balance"`
+	PointsEarned   float64    `json:"points_earned"`
+	PointsDeducted float64    `json:"points_deducted"`
+	PointsBalance  float64    `json:"points_balance"`
 	CustomerID     *string    `json:"customer_id"`
 	CustomerName   string     `json:"customer_name"`
 	CustomerPhone  string     `json:"customer_phone_number"`
